fix: accordion breakpoint example
