+++ src/routes/docs-examples/components/accordion/BpRange.svelte
@@ -2,22 +2,16 @@
   import { AccordionItem, useMediaQuery, useBreakpoints, P } from "flowbite-svelte";
   const breakpoints = useBreakpoints();
 
-  // Open from sm to lg (640px - 1023px)
-  const tabletRange = $derived(breakpoints.sm && !breakpoints.lg);
-
-  // Open on specific breakpoints only
-  const specificSizes = $derived((breakpoints.sm && !breakpoints.md) || (breakpoints.lg && !breakpoints.xl));
-
   // Custom pixel range
   const customRange = useMediaQuery("(min-width: 640px) and (max-width: 1023px)");
 </script>
 
-<AccordionItem open={tabletRange}>
+<AccordionItem open={breakpoints.sm && !breakpoints.lg}>
   {#snippet header()}Tablet Range (640px - 1023px){/snippet}
   <P>Open on tablets, closed on phones and large desktops.</P>
 </AccordionItem>
 
-<AccordionItem open={specificSizes}>
+<AccordionItem open={(breakpoints.sm && !breakpoints.md) || (breakpoints.lg && !breakpoints.xl)}>
   {#snippet header()}Small phones OR Large desktops only{/snippet}
   <P>Open on sm-only OR lg-only, closed on other sizes.</P>
 </AccordionItem>
