+++ static/themes/manifest.json
@@ -1,7 +1,1 @@
-[
-  "default",
-  "minimal",
-  "enterprise",
-  "playful",
-  "mono"
-]
+["default", "minimal", "enterprise", "playful", "mono"]
